fixing issues
